Code changes after Lint review

+++ libcore/src/com/samknows/measurement/activity/components/ProgressWheel.java
@@ -187,13 +187,13 @@ public class ProgressWheel extends View {
 	    
 	    textSize = (int) a.getDimension(R.styleable.ProgressWheel_textSize, textSize);
 	    
-	    textColor = (int) a.getColor(R.styleable.ProgressWheel_textColor, textColor);
+	    textColor = a.getColor(R.styleable.ProgressWheel_textColor, textColor);
 	    
 	    //setText(a.getString(R.styleable.ProgressWheel_title));
 	    
-	    rimColor = (int) a.getColor(R.styleable.ProgressWheel_rimColor, rimColor);
+	    rimColor = a.getColor(R.styleable.ProgressWheel_rimColor, rimColor);
 	    
-	    circleColor = (int) a.getColor(R.styleable.ProgressWheel_circleColor, circleColor);
+	    circleColor = a.getColor(R.styleable.ProgressWheel_circleColor, circleColor);
 	}
 
 	//----------------------------------
